refactor: modularize diagnostics endpoints
