--- pkgs/standards/autoapi/autoapi/v3/system/diagnostics.py
@@ -1,407 +1,0 @@
-# autoapi/v3/system/diagnostics.py
-"""
-Diagnostics endpoints for AutoAPI v3.
-
-Exposes a small router with:
-  - GET /healthz     → DB connectivity check (sync or async)
-  - GET /methodz     → list RPC-visible methods derived from OpSpecs
-  - GET /hookz       → per-op phase chains (sequential order)
-  - GET /planz       → runtime execution plan (flattened labels) per op
-
-Usage:
-    from autoapi.v3.system.diagnostics import mount_diagnostics
-    app.include_router(mount_diagnostics(api, get_db=get_db), prefix="/system")
-"""
-
-from __future__ import annotations
-
-import inspect
-import logging
-from types import SimpleNamespace
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Sequence,
-)
-
-try:
-    from ..types import Router, Request, Depends
-    from fastapi.responses import JSONResponse
-except Exception:  # pragma: no cover
-    # Lightweight shims so the module is importable without FastAPI
-    class Router:  # type: ignore
-        def __init__(self, *a, **kw):
-            self.routes = []
-
-        def add_api_route(
-            self, path: str, endpoint: Callable, methods: Sequence[str], **opts
-        ):
-            self.routes.append((path, methods, endpoint, opts))
-
-    class Request:  # type: ignore
-        def __init__(self, scope=None):
-            self.scope = scope or {}
-            self.state = SimpleNamespace()
-            self.query_params = {}
-
-    class JSONResponse(dict):  # type: ignore
-        def __init__(self, content: Any, status_code: int = 200):
-            super().__init__(content=content, status_code=status_code)
-
-
-from sqlalchemy import text
-from ..ops.types import PHASES
-from ..runtime.kernel import build_phase_chains
-from ..runtime import plan as _plan
-
-logger = logging.getLogger(__name__)
-
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Helpers
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-def _model_iter(api: Any) -> Iterable[type]:
-    models = getattr(api, "models", {}) or {}
-    return models.values()
-
-
-def _opspecs(model: type):
-    return getattr(getattr(model, "opspecs", SimpleNamespace()), "all", ()) or ()
-
-
-def _label_callable(fn: Any) -> str:
-    """Return a module-qualified label for *fn* suitable for hook listings."""
-    n = getattr(fn, "__qualname__", getattr(fn, "__name__", repr(fn)))
-    m = getattr(fn, "__module__", None)
-    return f"{m}.{n}" if m else n
-
-
-def _label_hook(fn: Any, phase: str) -> str:
-    subj = _label_callable(fn).replace(".", ":")
-    return f"hook:wire:{subj}@{phase}"
-
-
-async def _maybe_execute(db: Any, stmt: str):
-    try:
-        rv = db.execute(text(stmt))  # type: ignore[attr-defined]
-        if inspect.isawaitable(rv):
-            return await rv
-        return rv
-    except Exception:
-        # Some drivers prefer lowercase 'select 1'
-        rv = db.execute(text("select 1"))  # type: ignore[attr-defined]
-        if inspect.isawaitable(rv):
-            return await rv
-        return rv
-
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Builders
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-def _build_healthz_endpoint(dep: Optional[Callable[..., Any]]):
-    """
-    Returns a FastAPI endpoint function for /healthz.
-    If `dep` is provided, it's used as a dependency to supply `db`.
-    Otherwise, we try request.state.db.
-    """
-    if dep is not None:
-
-        async def _healthz(db: Any = Depends(dep)):
-            try:
-                await _maybe_execute(db, "SELECT 1")
-                return {"ok": True}
-            except Exception as e:  # pragma: no cover
-                logger.exception("/healthz failed")
-                return JSONResponse({"ok": False, "error": str(e)}, status_code=500)
-
-        return _healthz
-
-    async def _healthz(request: Request):
-        db = getattr(request.state, "db", None)
-        if db is None:
-            # No DB provided—treat as healthy from API process standpoint
-            return {"ok": True, "warning": "no-db"}
-        try:
-            await _maybe_execute(db, "SELECT 1")
-            return {"ok": True}
-        except Exception as e:  # pragma: no cover
-            logger.exception("/healthz failed")
-            return JSONResponse({"ok": False, "error": str(e)}, status_code=500)
-
-    return _healthz
-
-
-def _build_methodz_endpoint(api: Any):
-    cache: Optional[Dict[str, List[Dict[str, Any]]]] = None
-
-    async def _methodz():
-        nonlocal cache
-        """Ordered, canonical operation list."""
-        if cache is not None:
-            return cache
-
-        methods: List[Dict[str, Any]] = []
-        for model in _model_iter(api):
-            mname = getattr(model, "__name__", "Model")
-            for sp in _opspecs(model):
-                if not getattr(sp, "expose_rpc", True):
-                    continue
-                methods.append(
-                    {
-                        "method": f"{mname}.{sp.alias}",
-                        "model": mname,
-                        "alias": sp.alias,
-                        "target": sp.target,
-                        "arity": sp.arity,
-                        "persist": sp.persist,
-                        "request_model": getattr(sp, "request_model", None) is not None,
-                        "response_model": getattr(sp, "response_model", None)
-                        is not None,
-                        "routes": bool(getattr(sp, "expose_routes", True)),
-                        "rpc": bool(getattr(sp, "expose_rpc", True)),
-                        "tags": list(getattr(sp, "tags", ()) or (mname,)),
-                    }
-                )
-        methods.sort(key=lambda x: (x["model"], x["alias"]))
-        cache = {"methods": methods}
-        return cache
-
-    return _methodz
-
-
-def _build_hookz_endpoint(api: Any):
-    cache: Optional[Dict[str, Dict[str, Dict[str, List[str]]]]] = None
-
-    async def _hookz():
-        nonlocal cache
-        """
-        Expose hook execution order for each method.
-
-        Phases appear in runner order; error phases trail.
-        Within each phase, hooks are listed in execution order: global (None) hooks,
-        then method-specific hooks.
-        """
-        if cache is not None:
-            return cache
-
-        out: Dict[str, Dict[str, Dict[str, List[str]]]] = {}
-        for model in _model_iter(api):
-            mname = getattr(model, "__name__", "Model")
-            hooks_root = getattr(model, "hooks", SimpleNamespace())
-            # Build list of aliases from RPC ns or opspecs
-            alias_sources = set()
-            rpc_ns = getattr(model, "rpc", SimpleNamespace())
-            alias_sources.update(getattr(rpc_ns, "__dict__", {}).keys())
-            for sp in _opspecs(model):
-                alias_sources.add(sp.alias)
-
-            model_map: Dict[str, Dict[str, List[str]]] = {}
-            for alias in sorted(alias_sources):
-                alias_ns = getattr(hooks_root, alias, None) or SimpleNamespace()
-                phase_map: Dict[str, List[str]] = {}
-                for ph in PHASES:
-                    steps = list(getattr(alias_ns, ph, []) or [])
-                    if steps:
-                        phase_map[ph] = [_label_callable(fn) for fn in steps]
-                if phase_map:
-                    model_map[alias] = phase_map
-            if model_map:
-                out[mname] = model_map
-        cache = out
-        return cache
-
-    return _hookz
-
-
-def _build_planz_endpoint(api: Any):
-    cache: Optional[Dict[str, Dict[str, List[str]]]] = None
-
-    async def _planz():
-        nonlocal cache
-        """Expose the runtime step sequence for each operation."""
-        if cache is not None:
-            return cache
-        out: Dict[str, Dict[str, List[str]]] = {}
-        for model in _model_iter(api):
-            mname = getattr(model, "__name__", "Model")
-            model_map: Dict[str, List[str]] = {}
-            compiled_plan = getattr(
-                getattr(model, "runtime", SimpleNamespace()), "plan", None
-            )
-            if compiled_plan is None:
-                specs = getattr(model, "__autoapi_colspecs__", None) or getattr(
-                    model, "__autoapi_cols__", None
-                )
-                if specs is not None:
-                    try:
-                        compiled_plan = _plan.attach_atoms_for_model(model, specs)
-                    except Exception:
-                        compiled_plan = None
-            for sp in _opspecs(model):
-                seq: List[str] = []
-                persist = getattr(sp, "persist", "default") != "skip"
-                if compiled_plan is not None:
-                    deps: List[str] = [
-                        _label_callable(d) if callable(d) else str(d)
-                        for d in getattr(sp, "deps", []) or []
-                    ]
-                    secdeps: List[str] = [
-                        _label_callable(d) if callable(d) else str(d)
-                        for d in getattr(sp, "secdeps", []) or []
-                    ]
-                    hooks_root = getattr(model, "hooks", SimpleNamespace())
-                    alias_ns = getattr(hooks_root, sp.alias, None)
-                    hook_map: Dict[str, List[str]] = {}
-                    if alias_ns is not None:
-                        for ph in PHASES:
-                            steps = getattr(alias_ns, ph, []) or []
-                            if steps:
-                                hook_map[ph] = [_label_hook(fn, ph) for fn in steps]
-                    labels = _plan.flattened_order(
-                        compiled_plan,
-                        persist=persist,
-                        include_system_steps=True,
-                        deps=deps,
-                        secdeps=secdeps,
-                        hooks=hook_map,
-                    )
-                    labels = [
-                        lbl
-                        for lbl in labels
-                        if not (
-                            lbl.startswith("HANDLER:hook:")
-                            and "hook:sys:handler:crud@HANDLER" not in lbl
-                        )
-                    ]
-                    if sp.target == "custom" or getattr(sp, "persist", "default") in {
-                        "override"
-                    }:
-                        labels = [
-                            lbl
-                            for lbl in labels
-                            if "hook:sys:handler:crud@HANDLER" not in lbl
-                        ]
-                    seq = labels
-                else:
-                    deps: List[str] = [
-                        _label_callable(d) if callable(d) else str(d)
-                        for d in getattr(sp, "deps", []) or []
-                    ]
-                    secdeps: List[str] = [
-                        _label_callable(d) if callable(d) else str(d)
-                        for d in getattr(sp, "secdeps", []) or []
-                    ]
-                    chains = build_phase_chains(model, sp.alias)
-                    seq.extend(f"PRE_TX:secdep:{s}" for s in secdeps)
-                    seq.extend(f"PRE_TX:dep:{d}" for d in deps)
-                    for ph in PHASES:
-                        if ph == "PRE_TX":
-                            continue
-                        if ph == "START_TX" and persist:
-                            seq.append("START_TX:hook:sys:txn:begin@START_TX")
-                        if (
-                            ph == "HANDLER"
-                            and persist
-                            and sp.target != "custom"
-                            and getattr(sp, "persist", "default") not in {"override"}
-                        ):
-                            seq.append("HANDLER:hook:sys:handler:crud@HANDLER")
-                        for step in chains.get(ph, []) or []:
-                            name = getattr(step, "__name__", "")
-                            if name in {"start_tx", "end_tx"}:
-                                continue
-                            seq.append(f"{ph}:{_label_hook(step, ph)}")
-                        if ph == "END_TX" and persist:
-                            seq.append("END_TX:hook:sys:txn:commit@END_TX")
-                seen_wire = set()
-                dedup_seq: List[str] = []
-                for lbl in seq:
-                    if lbl.startswith("hook:wire:"):
-                        if lbl in seen_wire:
-                            continue
-                        seen_wire.add(lbl)
-                    dedup_seq.append(lbl)
-                model_map[sp.alias] = dedup_seq
-            if model_map:
-                out[mname] = model_map
-        cache = out
-        return cache
-
-    return _planz
-
-
-# ───────────────────────────────────────────────────────────────────────────────
-# Public factory
-# ───────────────────────────────────────────────────────────────────────────────
-
-
-def mount_diagnostics(
-    api: Any,
-    *,
-    get_db: Optional[Callable[..., Any]] = None,
-) -> Router:
-    """
-    Create & return a Router that exposes:
-      GET /healthz
-      GET /methodz
-      GET /hookz
-      GET /planz
-    """
-    router = Router()
-
-    dep = get_db
-
-    router.add_api_route(
-        "/healthz",
-        _build_healthz_endpoint(dep),
-        methods=["GET"],
-        name="healthz",
-        tags=["system"],
-        summary="Health",
-        description="Database connectivity check.",
-    )
-    router.add_api_route(
-        "/methodz",
-        _build_methodz_endpoint(api),
-        methods=["GET"],
-        name="methodz",
-        tags=["system"],
-        summary="Methods",
-        description="Ordered, canonical operation list.",
-    )
-    router.add_api_route(
-        "/hookz",
-        _build_hookz_endpoint(api),
-        methods=["GET"],
-        name="hookz",
-        tags=["system"],
-        summary="Hooks",
-        description=(
-            "Expose hook execution order for each method.\n\n"
-            "Phases appear in runner order; error phases trail.\n"
-            "Within each phase, hooks are listed in execution order: "
-            "global (None) hooks, then method-specific hooks."
-        ),
-    )
-    router.add_api_route(
-        "/planz",
-        _build_planz_endpoint(api),
-        methods=["GET"],
-        name="planz",
-        tags=["system"],
-        summary="Plan",
-        description="Flattened runtime execution plan per operation.",
-    )
-
-    return router
-
-
-__all__ = ["mount_diagnostics"]
